src/main/java/com/ecoinv/inventorymanagement/InventoryManager.java
package com.ecoinv.inventorymanagement;

import com.ecoinv.inventorymanagement.model.InventoryItem;

import java.util.List;

public class InventoryManager {
    // Constructor
    public InventoryManager() {
        // Initialization code here (e.g., setting up API clients, database connections)
    }

    // Method to analyze sales data
    public void analyzeSalesData() {
        // Fetch sales data from a database or an external API
        // Process the data to identify sales patterns and trends
        System.out.println("Analyzing sales data...");
        // This could involve complex data analysis and machine learning predictions
        // Consider using a library or service for ML predictions
    }

    // Method to adjust inventory levels based on analyzed data
    public void adjustInventoryLevels() {
        // Based on the analysis, determine which products need more stock and which have excess
        System.out.println("Adjusting inventory levels...");
        // Adjust inventory records in the database or through an inventory management API
        // This step might involve setting thresholds for reordering or reducing stock
    }

    // Method for automated reordering with supplier systems
    public void automateReordering(List<InventoryItem> currentInventory) {
        for (InventoryItem item : currentInventory) {
            // Check if the current stock is below the reorder threshold
            if (item.getCurrentStock() < item.getThreshold()) {
                int reorderQuantity = calculateReorderQuantity(item);

                // Simulate placing an order with the supplier
                placeOrderWithSupplier(item.getProductName(), reorderQuantity);

                // Optionally, update the inventory item to reflect the placed order
                // This might involve increasing the 'on order' stock level
                item.increaseOnOrderStock(reorderQuantity);

                System.out.println("Reordered " + reorderQuantity + " units of " + item.getProductName());
            }
        }
    }

    // Calculate the reorder quantity based on past sales data, minimum order quantities, etc.
    private int calculateReorderQuantity(InventoryItem item) {
        // Placeholder for actual calculation logic
        // For simplicity, we're just returning the reorder quantity specified in the item
        return item.getReorderQuantity();
    }

    // Simulate placing an order with the supplier
    private void placeOrderWithSupplier(String productName, int quantity) {
        // Placeholder for actual integration with supplier systems
        // This could involve sending a REST API request, an email, etc.
    }
}
