src/test/java/com/ecoinv/inventorymanagement/InventoryManagerTest.java
package com.ecoinv.inventorymanagement;

import com.ecoinv.MockOrderSystem;
import org.junit.After;
import org.junit.Before;
import org.junit.Test;

public class InventoryManagerTest {

    private InventoryManager inventoryManager;
    private MockOrderSystem mockOrderSystem = new MockOrderSystem();
    // Setup before each test
    @Before
    public void setUp() {
        inventoryManager = new InventoryManager();
        // Initialize your InventoryManager with any required data
    }

    // Test case for analyzing sales data
    @Test
    public void testAnalyzeSalesData() {
        // Call the method you want to test
        inventoryManager.analyzeSalesData();
        // Assert the expected outcomes using assertTrue, assertEquals, etc.
        // For example:
        // assertEquals("Expected result", inventoryManager.getSomeResult());
    }

    // Test case for adjusting inventory levels
    @Test
    public void testAdjustInventoryLevels() {
        // Setup specific conditions for the test
        // Call the method to test
        inventoryManager.adjustInventoryLevels();
        // Assert the expected outcomes
    }


    @Test
    public void testAutomateReordering() {
       /* InventoryManager inventoryManager = new InventoryManager();
        List<InventoryItem> currentInventory = Arrays.asList(
                new InventoryItem("Product1", 90, 100, 50), // Product name, current stock, reorder threshold, reorder quantity
                new InventoryItem("Product2", 110, 100, 50)
        );

        inventoryManager.automateReordering(currentInventory);
        System.out.println("Orders placed: " + mockOrderSystem.getOrders()); // Assuming getOrders returns a map or list of orders
        // Verify orders placed for items below threshold
        assertTrue("Order placed for Product1", mockOrderSystem.orderPlacedFor("Product1"));
        assertFalse("No order placed for Product2", mockOrderSystem.orderPlacedFor("Product2"));
        */

    }

    // Cleanup after each test
    @After
    public void tearDown() {
        // Clean up resources, if necessary
    }
}
